[refactor] usePayment 구현, 결제 코드 분리
